Vibe coding challenge
This is the first time I've ever done vibe coding but it is so problematic and so infuriating that the only proper use case I can find for this is documentation definitely will not be repeating this experiment unless I am forced to. We have an upcoming game jam on an engine that I've never programmed on in this language I know very little about and they have their own AI bot so this is our attempt to try to learn how to mix human with AI. This description is probably the longest description I've ever written for a commit. 1/10